moved style to separated dir
