+++ src/engraving/libmscore/ottava.cpp
@@ -21,7 +21,9 @@
  */
 
 #include "ottava.h"
-#include "style.h"
+
+#include "style/style.h"
+
 #include "system.h"
 #include "measure.h"
 #include "xml.h"
