+++ src/engraving/libmscore/vibrato.cpp
@@ -24,8 +24,8 @@
 #include <cmath>
 
 #include "translation.h"
+#include "style/style.h"
 
-#include "style.h"
 #include "system.h"
 #include "measure.h"
 #include "xml.h"
